Refactored to usage of  all modules from main.py
